tutorial_v3/requirements.txt
requests==2.31.0
gspread==6.2.0
google-auth==2.38.0  # Заменяем oauth2client
pybit==5.5.0
python-telegram-bot==20.6
python-dotenv==1.0.1  # Актуальная версия на момент 2024 года
google-auth-oauthlib
google-auth-httplib2
google-api-python-client



